@@ -60,7 +60,7 @@ public class CopyMoreAction extends SimpleCommand {
             case COPY_CITE_KEY -> copyCiteKey();
             case COPY_KEY_AND_TITLE -> copyKeyAndTitle();
             case COPY_KEY_AND_LINK -> copyKeyAndLink();
-            case COPY_DOI -> copyDoi();
+            case COPY_DOI, COPY_DOI_URL -> copyDoi();
             default -> LOGGER.info("Unknown copy command.");
         }
     }
@@ -121,12 +121,21 @@ public class CopyMoreAction extends SimpleCommand {
     private void copyDoi() {
         List<BibEntry> entries = stateManager.getSelectedEntries();
 
-        // Collect all non-null DOIs.
-        List<String> dois = entries.stream()
-                                   .filter(entry -> entry.getDOI().isPresent())
-                                   .map(entry -> entry.getDOI().get().getDOI())
-                                   .collect(Collectors.toList());
+        // Collect all non-null DOI or DOI urls
+        if (action == StandardActions.COPY_DOI_URL) {
+            copyDoiList(entries.stream()
+                    .filter(entry -> entry.getDOI().isPresent())
+                    .map(entry -> entry.getDOI().get().getURIAsASCIIString())
+                    .collect(Collectors.toList()), entries.size());
+        } else {
+            copyDoiList(entries.stream()
+                    .filter(entry -> entry.getDOI().isPresent())
+                    .map(entry -> entry.getDOI().get().getDOI())
+                    .collect(Collectors.toList()), entries.size());
+        }
+    }
 
+    private void copyDoiList(List<String> dois, int size) {
         if (dois.isEmpty()) {
             dialogService.notify(Localization.lang("None of the selected entries have DOIs."));
             return;
@@ -135,13 +144,13 @@ public class CopyMoreAction extends SimpleCommand {
         final String copiedDois = String.join(",", dois);
         clipBoardManager.setContent(copiedDois);
 
-        if (dois.size() == entries.size()) {
+        if (dois.size() == size) {
             // All entries had DOIs.
             dialogService.notify(Localization.lang("Copied '%0' to clipboard.",
                     JabRefDialogService.shortenDialogMessage(copiedDois)));
         } else {
             dialogService.notify(Localization.lang("Warning: %0 out of %1 entries have undefined DOIs.",
-                    Integer.toString(entries.size() - dois.size()), Integer.toString(entries.size())));
+                    Integer.toString(size - dois.size()), Integer.toString(size)));
         }
     }
 
